make it better

+++ app/Http/Controllers/DogController.php
@@ -4,7 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Dog;
-use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
+
 
 class DogController extends Controller
 {
@@ -35,6 +36,7 @@ class DogController extends Controller
             'name' => $request->name,
             'age' => $request->age,
             'breed' => $request->breed,
+            'user_id' => Auth::user()->id,
         ]);
 
         return redirect()->route('dogs.index');
@@ -53,9 +55,13 @@ class DogController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit($id)
     {
         $dog = Dog::find($id);
+
+        if (Auth::user()->id !== $dog->user_id) {
+            return redirect('/')->with('error', 'Unauthorized Page');
+        }
 
         return view('dogs.edit', compact('dog'));
     }
@@ -83,6 +89,9 @@ class DogController extends Controller
     {
         $dog = Dog::find($id);
 
+        if (Auth::user()->id !== $dog->user_id) {
+            return redirect('/')->with('error', 'Unauthorized Page');
+        }
         $dog->delete();
 
         return redirect()->route('dogs.index');
